email check just to get logics and syntaxs

+++ demo/src/main/java/com/example/demo/student/StudentService.java
@@ -33,6 +33,10 @@ public class StudentService {
     }
 
     public Boolean DeleteStudent(UUID id) {
+        Boolean b = studentRepository.existsById(id);
+        if(!b){
+            throw new IllegalStateException("student with id : "+ id + "does not exists");
+        }
         return studentRepository.deleteStudentById(id);
     }
 }
